requirements.txt
jsonlines
tqdm
pillow
pickle
base64